More Job Info
Main change is adding more information for the job positions.
Added a progress animation while loading the data file.
Added a jumbotron to indicate that no result were found when using the
search box.
Updated the merge script to convert all location names to lower case.
This help with displaying the address for the job location.